perf: stop emitting _amount / _round in Contribute
two reasons for amount:
1. it's not actually needed. you can get the value from an event in the
subgraph from event.transaction.value
2. it was the compressed uint80 version of the amount.
that makes it harder to make tests around.
we can do the bit shifting in the subgraph independently.

in the case of round:
I'm not sure why it was emitted in Slot Curate...
but it seems redundant, surely the subgraph can infer the
nextRound of a dispute just by fetching currentRound and
adding one.

saves 750 gas