feat: adoption. neglected items can be adopted
this is a griefing protection mechanism to ensure
a griefer submitter cannot just go and "delete" an
item, forcing other submitters to wait until the item is
removed and submitting it.

- it would be expensive (repeating ipfs data)
- it would cause the item to be absent from the list
even if for a short amount of time
- it would lose the "history" of the item, because of
the newly generated evidenceGroupId of the resubmitted
item.

also, because owning the item is a liability, and this method
is always available as a forced transfer, it doesn't matter if
it creates "a race" to adopt an item, since adopting it
cannot cause any damage.