perf: stop emitting accountId for creation
since the subgraph acts as source of truth, no need
to add redundancy to the events.
also make account events emit the Cint32 stake
instead. the subgraph will always keep track of it
uncompressed, and will decompress it natively.

the reason you want to decompress natively is that
you want to get the real, exact amounts that are
living in the contract.