feat(contracts): challengeWindow
the proper solution to the edition
and challenge frontrunning dilema.
the challenger has to reference an
edition by using a timestamp. the
contract reverts if the timestamp is
too old. if it passes, then the subgraph
derives the edition from this timestamp
and it will be the edition displayed to
jurors and the main reference of the
dispute.