feat: Withdrawing as item state
as a way to discourage users starting
withdrawal period at "no cost", a way is
to signal the item in a different way that's
not quite inclusion. so, even if the items are
not young, they can be reclaimed by others.

on-chain consumers can reject this status,
and "canonical" inclusion won't consider them.

despite this, they're still challengeable