feat: utilize erc20 stakes
caveats, apart from removing some
dupe code in rule, an issue that matters is
that, in the current state, either a malicious
arbitrator or a faulty erc20 can result in an
item getting stuck forever.

some lists might have conflicts around this,
although, if this was the case, they should have
just used non-faulty erc20s or arbitrators instead.