wip: recentlykept
this feature will be removed soon.
it was needed to avoid someone from frontrunning
an onchain consumption. but since you will
be able to hold multiple, parallel disputes,
the item could be securely removed even if
there are purposely wrong challenges.

this will also mean that edits during a challenge
should be fine, as the challenge will still be ruled
on a particular edition.

this commit is made to salvage other things such as
a few comments.