fix: retracting
stopRetraction is not needed, you can just
recommit.
the ItemState "Retracting" is Dubious. not
needed internally for Stake Curate and it
would add extra view complexity. should
it just be removed? then protocols that
somehow care about it, can read it. but,
Retractions are pretty much a feature
intended for internal usage.