feat: maxStake
prevents some bankrun attacks in case
someone stakes a crazy amount that prevents
users from challenging.

you don't need to add an "Overcollateralized"
because the only way for the maxStake to be
changed is updating the list, and then it
becomes Outdated anyway.

remember that the 2nd bankrun protection
(autostakes) will require doing checks to
prevent adding too much on the item.