feat: nextStake, frontrunning protection
since items have variable stake, it could
happen that an item owner frontruns
a challenge reveal that is aimed towards
their item, and gets the stake collateralizing
it to be larger than what the tokens the
challenger passes can cover.

this update prevents that from happening,
as stake increases take some time to process.