fix: adopter no longer held liable for past
challengeWindow ended up being important.
previously, anyone who revives an item could
be held liable by a well timed challenge on
the previous edition, that could very well be
wrong (and precisely, the reason the item got
deleted in the first place).
this is solved by keeping track of the last time
an item was "rescued", as in, came from being
unincluded to being included. challenger
gets a full refund upon reveal if that's the case.
this should seldom occur in practice, as challengers
should just reveal quickly.

also fixed a bug in which Removed items were
not triggering burns on reveal. rationale was
given as a comment. to be fair, this has become
the case only after the removal of
arbitrator allowance.