refactor: remove arbitrator & party from evidence
saves calldata, removing the need to pass
an arbitrator. it was just kept before for
erc-1497 compatibility, but the history of
the item is not related to arbitrators.

the way it will work is, the frontend will render
this evidence, and the evidence display will
provide a way to the arbitrator to read the
evidence as well.