docs: short description of account view
showing pretty stats and information on a
particular stake curate account is nice to have,
and should be there on the final version,
but it will be better to worry about extras
after the frontend can handle the basics.

the minimal version requires withdrawing
stakes somehow, and it should be on this view.