fix: check list legality
listLegalCheck is back. it is only
checked on createList and updateList,
so it stays gas friendly.
three constants related to lists being possible weren't being
checked.

another way to deal with this would be to create a "legalizeList"
method that just goes through a list and enforces min/max settings
this could be done if general settings ever need to be updated