refactor: remove refreshItem
there was an edge case in which the attacker frontruns an
edit to an item to make it wrong, so that another party refreshes
it and gets stuck with the bad edition, and then they get challenged.

a solution to this was passing the timestamp of the edition that was
to be refreshed. but at this point, this function is much too similar
to edit item. it works in a similar way, it just doesn;t
create a new edition. why not create this new edition and
be done with it? if this ever becomes too expensive,
this can still be upgraded by proxy and have refresh back.