feat: check item version when edit item
for lists in which the previous editions and the
evolution in the fields of item can affect the
acceptance of the current edition, some frontrunning
attacks might be possible.
having this safe check prevents having to worry about
this aspect and having to remember that it is a limitation.

why does it have to be added? just like the
argument _forListVersion, it is added because without
account abstraction, a regular wallet cannot batch this
in a list of transactions to verify the listVersion is the correct one.

when account abstraction is over, stake curate can be
proxy upgraded and these two timestamps removed to save calldata.