feat: bigger burns from disputes
disputes need to be more taxing than just failing a commit
this is because malicious item submitters could, purposely,
prepare a commit right before causing some damage
(for example, submitting troll messages in a telegram chat)

and challenge themselves right after. this would allow them
to cause some degree of damage for MIN_TIME_FOR_REVEAL time
for just the BURN_RATE + arbFees,
which, in the case of failed commits, can be low.